Rails 3.2.2 compatiblity.
One of our reference data types has a code of VALID, which was
generating a valid? method that overwrote ActiveRecord's version. By
telling ActiveRecord we don't need validations when pulling over data
from the development environment, we side step that issue.

I still need to figure out what the proper behavior is for cases like
this. Perhaps throw an exception.