Removing dynamic loading hook logic.
I had added the dynamic loading hook logic because we used to load
reference data from fixtures, and we needed to delay the loading of the
reference data until after the fixtures had loaded. Otherwise, no
reference data would be found in the database.

Since we now have a rake task that copies the reference data over from
the development database, we can safely load all the reference data
objects when a class is marked as reference data. This removes a big
chunk of the magic.