Removing distinction between permanent and volatile attributes.
We've only had one use case where we need to reload the cached data
between requests because there is UI to edit the underlying rows. I plan
to implement a method for reloading the reference data objects, so we
can kick this detail to the application to manage through some sort of
invaliations scheme (say incrementing a memcache key, for example).