Provide a way for classes to opt out of auto-fixture behavior.
Normally, all reference data from development is copied to the test
database before the tests run, and all the primary keys are changed to
follow the pretty fixtures logic. With that behavior, you don't need to
maintain a parallel set of fixtures for all your reference data.

This commits adds a per-class flag that opts that class out of that
behavior.