Adding back some support for delayed loading and reloading the data.
Previously, we loaded the reference data from the database when
acts_as_reference_data was called. This causes errors if the reference
data classes are define during a db:migrate:reset call in Rails.

To work around that issue, we'll try to load the data from the database
in that call, but if that fails, we'll continue on. The railtie will
call ActsAsReferenceData.load_reference_data! to go through all the
classes and ensure that they are fully loaded.

Additionally, in migrations, it is sometimes useful to reload the
reference data. For example, you may have a migration that define a new
data type, and a later migration that uses that type. To solve for those
caching issues, we'll bring back reload_reference_data to reload the
objects from the database.